+++ src/referentielijsten/conf/base.py
@@ -1,355 +1,38 @@
-import os
-
 from django.urls import reverse_lazy
 
 from import_export.formats.base_formats import DEFAULT_FORMATS
+from notifications_api_common.settings import *  # noqa
+from open_api_framework.conf.base import *  # noqa
 
 from .api import *  # noqa
-from .utils import config
-
-# Build paths inside the project, so further paths can be defined relative to
-# the code root.
-
-DJANGO_PROJECT_DIR = os.path.abspath(
-    os.path.join(os.path.dirname(__file__), os.path.pardir)
-)
-
-BASE_DIR = os.path.abspath(
-    os.path.join(DJANGO_PROJECT_DIR, os.path.pardir, os.path.pardir)
-)
 
 #
 # Core Django settings
 #
-# SITE_ID = config("SITE_ID", default=1)
-
-# SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = config("SECRET_KEY")
-
-# NEVER run with DEBUG=True in production-like environments
-DEBUG = config("DEBUG", default=False)
-
-# = domains we're running on
-ALLOWED_HOSTS = config("ALLOWED_HOSTS", default="", split=True)
-USE_X_FORWARDED_HOST = config("USE_X_FORWARDED_HOST", default=False)
-
-IS_HTTPS = config("IS_HTTPS", default=not DEBUG)
-
-# Internationalization
-# https://docs.djangoproject.com/en/2.0/topics/i18n/
-
-LANGUAGE_CODE = "nl"
-
-TIME_ZONE = "UTC"
-
-USE_I18N = True
-
-USE_L10N = True
-
-USE_TZ = True
-
-USE_THOUSAND_SEPARATOR = True
-
-#
-# DATABASE and CACHING setup
-#
-DATABASES = {
-    "default": {
-        "ENGINE": config("DB_ENGINE", "django.db.backends.postgresql"),
-        "NAME": config("DB_NAME", "referentielijsten"),
-        "USER": config("DB_USER", "referentielijsten"),
-        "PASSWORD": config("DB_PASSWORD", "referentielijsten"),
-        "HOST": config("DB_HOST", "localhost"),
-        "PORT": config("DB_PORT", 5432),
-    }
-}
-
-DEFAULT_AUTO_FIELD = "django.db.models.AutoField"
-
-CACHES = {
-    "default": {
-        "BACKEND": "django_redis.cache.RedisCache",
-        "LOCATION": f"redis://{config('CACHE_DEFAULT', 'localhost:6379/0')}",
-        "OPTIONS": {
-            "CLIENT_CLASS": "django_redis.client.DefaultClient",
-            "IGNORE_EXCEPTIONS": True,
-        },
-    },
-    "axes": {
-        "BACKEND": "django_redis.cache.RedisCache",
-        "LOCATION": f"redis://{config('CACHE_AXES', 'localhost:6379/0')}",
-        "OPTIONS": {
-            "CLIENT_CLASS": "django_redis.client.DefaultClient",
-            "IGNORE_EXCEPTIONS": True,
-        },
-    },
-}
-
 
 #
 # APPLICATIONS enabled for this project
 #
-
-INSTALLED_APPS = [
-    "django.contrib.auth",
-    "django.contrib.sessions",
-    "django.contrib.contenttypes",
-    # NOTE: If enabled, at least one Site object is required and
-    # uncomment SITE_ID above.
-    # 'django.contrib.sites',
-    "django.contrib.messages",
-    "django.contrib.staticfiles",
-    # Two-factor authentication in the Django admin, enforced.
-    "django_otp",
-    "django_otp.plugins.otp_static",
-    "django_otp.plugins.otp_totp",
-    "two_factor",
-    "two_factor.plugins.webauthn",  # USB key/token support
-    "maykin_2fa",
-    # Optional applications.
-    "ordered_model",
-    "django_admin_index",
-    "django.contrib.admin",
-    # External applications.
-    "axes",
-    "mozilla_django_oidc",
-    "mozilla_django_oidc_db",
-    "drf_spectacular",
-    "rest_framework",
-    "solo",
-    "django_jsonform",
-    "vng_api_common",
+INSTALLED_APPS = INSTALLED_APPS + [
     "import_export",
     # Project applications.
     "referentielijsten.accounts",
     "referentielijsten.utils",
     "referentielijsten.api",
 ]
-
-MIDDLEWARE = [
-    "django.middleware.security.SecurityMiddleware",
-    "django.contrib.sessions.middleware.SessionMiddleware",
-    # 'django.middleware.locale.LocaleMiddleware',
-    "django.middleware.common.CommonMiddleware",
-    "django.middleware.csrf.CsrfViewMiddleware",
-    "django.contrib.auth.middleware.AuthenticationMiddleware",
-    "mozilla_django_oidc_db.middleware.SessionRefresh",
-    "maykin_2fa.middleware.OTPMiddleware",
-    "django.contrib.messages.middleware.MessageMiddleware",
-    "django.middleware.clickjacking.XFrameOptionsMiddleware",
-    # should be last according to docs
-    "axes.middleware.AxesMiddleware",
-]
-
-ROOT_URLCONF = "referentielijsten.urls"
-
-# List of callables that know how to import templates from various sources.
-TEMPLATE_LOADERS = (
-    "django.template.loaders.filesystem.Loader",
-    "django.template.loaders.app_directories.Loader",
-)
-
-TEMPLATES = [
-    {
-        "BACKEND": "django.template.backends.django.DjangoTemplates",
-        "DIRS": [os.path.join(DJANGO_PROJECT_DIR, "templates")],
-        "APP_DIRS": False,  # conflicts with explicity specifying the loaders
-        "OPTIONS": {
-            "context_processors": [
-                "django.template.context_processors.debug",
-                "django.template.context_processors.request",
-                "django.contrib.auth.context_processors.auth",
-                "django.contrib.messages.context_processors.messages",
-                "referentielijsten.utils.context_processors.settings",
-            ],
-            "loaders": TEMPLATE_LOADERS,
-        },
-    },
-]
-
-WSGI_APPLICATION = "referentielijsten.wsgi.application"
-
-# Translations
-LOCALE_PATHS = (os.path.join(DJANGO_PROJECT_DIR, "conf", "locale"),)
-
-#
-# SERVING of static and media files
-#
-
-STATIC_URL = "/static/"
-
-STATIC_ROOT = os.path.join(BASE_DIR, "static")
-
-# Additional locations of static files
-STATICFILES_DIRS = [os.path.join(DJANGO_PROJECT_DIR, "static")]
-
-# List of finder classes that know how to find static files in
-# various locations.
-STATICFILES_FINDERS = [
-    "django.contrib.staticfiles.finders.FileSystemFinder",
-    "django.contrib.staticfiles.finders.AppDirectoriesFinder",
-]
-
-MEDIA_ROOT = os.path.join(BASE_DIR, "media")
-
-MEDIA_URL = "/media/"
-
-FILE_UPLOAD_PERMISSIONS = 0o644
-
-#
-# Sending EMAIL
-#
-EMAIL_HOST = config("EMAIL_HOST", default="localhost")
-EMAIL_PORT = config(
-    "EMAIL_PORT", default=25
-)  # disabled on Google Cloud, use 487 instead
-EMAIL_HOST_USER = config("EMAIL_HOST_USER", default="")
-EMAIL_HOST_PASSWORD = config("EMAIL_HOST_PASSWORD", default="")
-EMAIL_USE_TLS = config("EMAIL_USE_TLS", default=False)
-EMAIL_TIMEOUT = 10
-
-DEFAULT_FROM_EMAIL = "referentielijsten@example.com"
-
-#
-# LOGGING
-#
-LOG_STDOUT = config("LOG_STDOUT", default=False)
-
-LOGGING_DIR = os.path.join(BASE_DIR, "log")
-
-LOGGING = {
-    "version": 1,
-    "disable_existing_loggers": False,
-    "formatters": {
-        "verbose": {
-            "format": "%(asctime)s %(levelname)s %(name)s %(module)s %(process)d %(thread)d  %(message)s"
-        },
-        "timestamped": {"format": "%(asctime)s %(levelname)s %(name)s  %(message)s"},
-        "simple": {"format": "%(levelname)s  %(message)s"},
-        "performance": {
-            "format": "%(asctime)s %(process)d | %(thread)d | %(message)s",
-        },
-    },
-    "filters": {
-        "require_debug_false": {"()": "django.utils.log.RequireDebugFalse"},
-    },
-    "handlers": {
-        "mail_admins": {
-            "level": "ERROR",
-            "filters": ["require_debug_false"],
-            "class": "django.utils.log.AdminEmailHandler",
-        },
-        "null": {
-            "level": "DEBUG",
-            "class": "logging.NullHandler",
-        },
-        "console": {
-            "level": "DEBUG",
-            "class": "logging.StreamHandler",
-            "formatter": "timestamped",
-        },
-        "django": {
-            "level": "DEBUG",
-            "class": "logging.handlers.RotatingFileHandler",
-            "filename": os.path.join(LOGGING_DIR, "django.log"),
-            "formatter": "verbose",
-            "maxBytes": 1024 * 1024 * 10,  # 10 MB
-            "backupCount": 10,
-        },
-        "project": {
-            "level": "DEBUG",
-            "class": "logging.handlers.RotatingFileHandler",
-            "filename": os.path.join(LOGGING_DIR, "referentielijsten.log"),
-            "formatter": "verbose",
-            "maxBytes": 1024 * 1024 * 10,  # 10 MB
-            "backupCount": 10,
-        },
-        "performance": {
-            "level": "INFO",
-            "class": "logging.handlers.RotatingFileHandler",
-            "filename": os.path.join(LOGGING_DIR, "performance.log"),
-            "formatter": "performance",
-            "maxBytes": 1024 * 1024 * 10,  # 10 MB
-            "backupCount": 10,
-        },
-    },
-    "loggers": {
-        "referentielijsten": {
-            "handlers": ["project"] if not LOG_STDOUT else ["console"],
-            "level": "INFO",
-            "propagate": True,
-        },
-        "django.request": {
-            "handlers": ["django"] if not LOG_STDOUT else ["console"],
-            "level": "ERROR",
-            "propagate": True,
-        },
-        "django.template": {
-            "handlers": ["console"],
-            "level": "INFO",
-            "propagate": True,
-        },
-        "mozilla_django_oidc": {
-            "handlers": ["project"] if not LOG_STDOUT else ["console"],
-            "level": "DEBUG",
-        },
-    },
-}
-
-#
-# AUTH settings - user accounts, passwords, backends...
-#
-AUTH_USER_MODEL = "accounts.User"
-
-AUTH_PASSWORD_VALIDATORS = [
-    {
-        "NAME": "django.contrib.auth.password_validation.UserAttributeSimilarityValidator"
-    },
-    {"NAME": "django.contrib.auth.password_validation.MinimumLengthValidator"},
-    {"NAME": "django.contrib.auth.password_validation.CommonPasswordValidator"},
-    {"NAME": "django.contrib.auth.password_validation.NumericPasswordValidator"},
-]
-
-
-# Allow logging in with both username+password and email+password
-AUTHENTICATION_BACKENDS = [
-    "axes.backends.AxesBackend",
-    "referentielijsten.accounts.backends.UserModelEmailBackend",
-    "django.contrib.auth.backends.ModelBackend",
-    "mozilla_django_oidc_db.backends.OIDCAuthenticationBackend",
-]
-
-SESSION_COOKIE_NAME = "referentielijsten_sessionid"
-SESSION_ENGINE = "django.contrib.sessions.backends.cache"
-
-LOGIN_URL = reverse_lazy("admin:login")
-LOGIN_REDIRECT_URL = reverse_lazy("admin:index")
-LOGOUT_REDIRECT_URL = reverse_lazy("admin:index")
+# open-api-framework currently installs `django.contrib.sites`, but we want to move
+# away from this in the future
+INSTALLED_APPS.remove("django.contrib.sites")
 
 #
 # SECURITY settings
 #
-SESSION_COOKIE_SECURE = IS_HTTPS
-SESSION_COOKIE_HTTPONLY = True
-
-CSRF_COOKIE_SECURE = IS_HTTPS
 CSRF_FAILURE_VIEW = "referentielijsten.accounts.views.csrf_failure"
-
-X_FRAME_OPTIONS = "DENY"
-
-#
-# FIXTURES
-#
-
-FIXTURE_DIRS = (os.path.join(DJANGO_PROJECT_DIR, "fixtures"),)
 
 #
 # Custom settings
 #
 PROJECT_NAME = "Referentielijsten"
-ENVIRONMENT = config("ENVIRONMENT", "")
-SHOW_ALERT = True
-ENABLE_ADMIN_NAV_SIDEBAR = config("ENABLE_ADMIN_NAV_SIDEBAR", default=False)
 
 # This setting is used by the csrf_failure view (accounts app).
 # You can specify any path that should match the request.path
@@ -357,26 +40,6 @@ ENABLE_ADMIN_NAV_SIDEBAR = config("ENABLE_ADMIN_NAV_SIDEBAR", default=False)
 # multiple login urls defined.
 LOGIN_URLS = [reverse_lazy("admin:login")]
 
-if "GIT_SHA" in os.environ:
-    GIT_SHA = config("GIT_SHA", "")
-# in docker (build) context, there is no .git directory
-elif os.path.exists(os.path.join(BASE_DIR, ".git")):
-    try:
-        import git
-    except ImportError:
-        GIT_SHA = None
-    else:
-        repo = git.Repo(search_parent_directories=True)
-        try:
-            GIT_SHA = repo.head.object.hexsha
-        except (
-            ValueError
-        ):  # on startproject initial runs before any git commits have been made
-            GIT_SHA = repo.active_branch.name
-else:
-    GIT_SHA = None
-
-RELEASE = config("RELEASE", GIT_SHA)
 
 # Default (connection timeout, read timeout) for the requests library (in seconds)
 REQUESTS_DEFAULT_TIMEOUT = (10, 30)
@@ -390,44 +53,8 @@ REQUESTS_DEFAULT_TIMEOUT = (10, 30)
 #
 # Django-Admin-Index
 #
-ADMIN_INDEX_SHOW_REMAINING_APPS = False
-ADMIN_INDEX_AUTO_CREATE_APP_GROUP = False
-ADMIN_INDEX_SHOW_REMAINING_APPS_TO_SUPERUSERS = True
 ADMIN_INDEX_DISPLAY_DROP_DOWN_MENU_CONDITION_FUNCTION = (
     "referentielijsten.utils.django_two_factor_auth.should_display_dropdown_menu"
-)
-
-#
-# DJANGO-AXES
-#
-AXES_CACHE = "axes"  # refers to CACHES setting
-# The number of login attempts allowed before a record is created for the
-# failed logins. Default: 3
-AXES_FAILURE_LIMIT = 10
-# If set, defines a period of inactivity after which old failed login attempts
-# will be forgotten. Can be set to a python timedelta object or an integer. If
-# an integer, will be interpreted as a number of hours. Default: None
-AXES_COOLOFF_TIME = 1
-# If set, specifies a template to render when a user is locked out. Template
-# receives cooloff_time and failure_limit as context variables. Default: None
-AXES_LOCKOUT_TEMPLATE = "account_blocked.html"
-AXES_LOCKOUT_PARAMETERS = [["ip_address", "user_agent", "username"]]
-# By default, Axes obfuscates values for formfields named "password", but the admin
-# interface login formfield name is "auth-password", so we obfuscate that as well
-AXES_SENSITIVE_PARAMETERS = ["password", "auth-password"]  # nosec
-
-# The default meta precedence order
-IPWARE_META_PRECEDENCE_ORDER = (
-    "HTTP_X_FORWARDED_FOR",
-    "X_FORWARDED_FOR",  # <client>, <proxy1>, <proxy2>
-    "HTTP_CLIENT_IP",
-    "HTTP_X_REAL_IP",
-    "HTTP_X_FORWARDED",
-    "HTTP_X_CLUSTER_CLIENT_IP",
-    "HTTP_FORWARDED_FOR",
-    "HTTP_FORWARDED",
-    "HTTP_VIA",
-    "REMOTE_ADDR",
 )
 
 #
@@ -437,42 +64,10 @@ IPWARE_META_PRECEDENCE_ORDER = (
 # those settings too.
 #
 # we run the admin site monkeypatch instead.
-TWO_FACTOR_PATCH_ADMIN = False
 # Relying Party name for WebAuthn (hardware tokens)
 TWO_FACTOR_WEBAUTHN_RP_NAME = "referentielijsten"
-# use platform for fingerprint readers etc., or remove the setting to allow any.
-# cross-platform would limit the options to devices like phones/yubikeys
-TWO_FACTOR_WEBAUTHN_AUTHENTICATOR_ATTACHMENT = "cross-platform"
-# add entries from AUTHENTICATION_BACKENDS that already enforce their own two-factor
-# auth, avoiding having some set up MFA again in the project.
-MAYKIN_2FA_ALLOW_MFA_BYPASS_BACKENDS = [
-    # "mozilla_django_oidc_db.backends.OIDCAuthenticationBackend",
-]
-
-# Elastic APM
-ELASTIC_APM_SERVER_URL = os.getenv("ELASTIC_APM_SERVER_URL", None)
-ELASTIC_APM = {
-    "SERVICE_NAME": f"referentielijsten {ENVIRONMENT}",
-    "SECRET_TOKEN": config("ELASTIC_APM_SECRET_TOKEN", "default"),
-    "SERVER_URL": ELASTIC_APM_SERVER_URL,
-}
-if not ELASTIC_APM_SERVER_URL:
-    ELASTIC_APM["ENABLED"] = False
-    ELASTIC_APM["SERVER_URL"] = "http://localhost:8200"
 
 #
-# Mozilla Django OIDC DB settings
+# django-import-export
 #
-OIDC_AUTHENTICATE_CLASS = "mozilla_django_oidc_db.views.OIDCAuthenticationRequestView"
-OIDC_CALLBACK_CLASS = "mozilla_django_oidc_db.views.OIDCCallbackView"
-MOZILLA_DJANGO_OIDC_DB_CACHE = "oidc"
-MOZILLA_DJANGO_OIDC_DB_CACHE_TIMEOUT = 5 * 60
-
-# Subpath (optional)
-# This environment variable can be configured during deployment.
-SUBPATH = config("SUBPATH", None)
-if SUBPATH:
-    SUBPATH = f"/{SUBPATH.strip('/')}"
-
-
 IMPORT_EXPORT_FORMATS = DEFAULT_FORMATS