+++ CHANGELOG.rst
@@ -2,6 +2,20 @@
 Change history
 ==============
 
+0.2.0
+=====
+
+*?? ??, 2024*
+
+.. warning::
+
+   Deployment tooling updates required - additional containers needed.
+
+   Redis is now required as a cache backend, make sure to add and configure a Redis container
+
+.. warning::
+
+   Two factor authentication was added (by default it is enabled, to disable it, set the ``DISABLE_2FA`` envvar to ``True``)
 
 0.1.0
 =====